// src/apis/models/CompanyFinancialDetails.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import { Amount, AmountFromJSON, AmountFromJSONTyped, AmountToJSON } from './Amount';
import { Tag, TagFromJSON, TagFromJSONTyped, TagToJSON } from './Tag';

/**
 *
 * @export
 * @interface CompanyFinancialDetails
 */
export interface CompanyFinancialDetails {
  /**
   *
   * @type {Amount}
   * @memberof CompanyFinancialDetails
   */
  expectedTransactionAmountPerMonth?: Amount;
  /**
   *
   * @type {Amount}
   * @memberof CompanyFinancialDetails
   */
  expectedTurnoverPerMonth?: Amount;
  /**
   * Additional information that can be added via tags
   * @type {Array<Tag>}
   * @memberof CompanyFinancialDetails
   */
  tags?: Array<Tag>;
}

export function CompanyFinancialDetailsFromJSON(json: any): CompanyFinancialDetails {
  return CompanyFinancialDetailsFromJSONTyped(json, false);
}

export function CompanyFinancialDetailsFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): CompanyFinancialDetails {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    expectedTransactionAmountPerMonth: !exists(json, 'expectedTransactionAmountPerMonth')
      ? undefined
      : AmountFromJSON(json['expectedTransactionAmountPerMonth']),
    expectedTurnoverPerMonth: !exists(json, 'expectedTurnoverPerMonth')
      ? undefined
      : AmountFromJSON(json['expectedTurnoverPerMonth']),
    tags: !exists(json, 'tags') ? undefined : (json['tags'] as Array<any>).map(TagFromJSON),
  };
}

export function CompanyFinancialDetailsToJSON(value?: CompanyFinancialDetails | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    expectedTransactionAmountPerMonth: AmountToJSON(value.expectedTransactionAmountPerMonth),
    expectedTurnoverPerMonth: AmountToJSON(value.expectedTurnoverPerMonth),
    tags: value.tags === undefined ? undefined : (value.tags as Array<any>).map(TagToJSON),
  };
}