// src/apis/models/Tag.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * Generic key-value pair model to append a custom variable being sent to Flagright API
 * @export
 * @interface Tag
 */
export interface Tag {
  /**
   * Key value when you are creating a custom variable
   * @type {string}
   * @memberof Tag
   */
  key: string;
  /**
   * Value for a given key when you are creating a custom variable
   * @type {string}
   * @memberof Tag
   */
  value: string;
}

export function TagFromJSON(json: any): Tag {
  return TagFromJSONTyped(json, false);
}

export function TagFromJSONTyped(json: any, ignoreDiscriminator: boolean): Tag {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    key: json['key'],
    value: json['value'],
  };
}

export function TagToJSON(value?: Tag | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    key: value.key,
    value: value.value,
  };
}